--- src/Components/Media_Container/useAudio.js
@@ -1,7 +1,0 @@
-import React from "react";
-
-const useAudio = () => {
-  return <div></div>;
-};
-
-export default useAudio;
